+++ item-manager/src/main/java/com/example/item_manager/service/CompanyService.java
@@ -25,7 +25,7 @@ public class CompanyService {
         return CompanyRepository.findAll();
     }
 
-    public Optional<Company> getCompanyById(Long id) {
+    public Optional<Company> getCompanyById(int id) {
         return CompanyRepository.findById(id);
     }
     
@@ -37,7 +37,7 @@ public class CompanyService {
         return CompanyRepository.findCompanyByLocation(location);
     }
 
-    public Company updateCompany(Long id, Company CompanyDetails) {
+    public Company updateCompany(int id, Company CompanyDetails) {
         Company Company = CompanyRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Item not found"));
         Company.setName(CompanyDetails.getName());
@@ -45,7 +45,7 @@ public class CompanyService {
         return CompanyRepository.save(Company);
     }
 
-    public void deleteCompany(Long id) {
+    public void deleteCompany(int id) {
         Company Company = CompanyRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Item not found"));
         CompanyRepository.delete(Company);
